Lab 6 done

+++ roblesgonzalez.guillermo/I6/cosx.h
@@ -1,5 +1,5 @@
 #ifndef SPBSPU_LABS_2022_AAP_904_A_COSX_H
 #define SPBSPU_LABS_2022_AAP_904_A_COSX_H
 #include <cstddef>
-double cosx(double x, double error, unsigned int cos_members);
+double cosx(double x, double error, size_t cos_members);
 #endif
